FUCKING diffball's free crapped out my run against oo. bastard, I'm fixing that soon... not even sure why the bitch segfaults, but she does... turned down multipass from starting at 512 to starting at 128.  possibly result's in a less optimal, but betting the correcting nature of onehalfpass will handle that fine. Unless a source has had *very* little changes, finding a 512|256 is a bit hard, going by my tests w/ oo. Make it dynamic I guess, prob best method.
Fixed a bug in dcbuffer.c for LLM, where no commands registered (usually
brought about by a bug in hash.c or diff-algs.c) would result in an
incorrect delta output, basically outputting no commands.
Fixed that, now it'll register that the entire file needs to be added as
it should for no copy commands.

Added a mild kludge to hash.c, basically verying the start position.
The reason for this is that a rhash of MOD_HASH type has a variable
starting position *always*, since it's not skipping every sample_rate
bytes, it's going through till it can find an insert, then skips to the
next sample_rate window.
SORT_HASH, by it's nature (not being neutered by the modulus op),
doesn't have a variable start.
It does now, albeit crappy.  it starts at seed_len *.5... did this so
that it's not always checking for the same prefix on it's blocks...
crappy, but it ought to work for the moment till I write something
better (reorganzing and chucking dups at certain intervals, and updating
the sample_rate to reflect for instance).

Side note, sample_rate neuters the hash_size slightly as one goes down
in seed size, mainly since duplicates become more common.  Annoying, but
the above hack (reorganizing, cleansing dups, and updating sample_rate)
ought to be a decent way to handle it for now.

bdelta write support appears to have a bug, conversion of a patch and
then trying to patch it fails on an assert, ver_pos < size1.  Not sure
why, haven't looked to fix it yet.